+++ BBIP/BBIP/Feature/Posting/PostingDetailViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  PostingDetailViewModel.swift
+//  BBIP
+//
+//  Created by 이건우 on 9/30/24.
+//
+
+import Foundation
